enrichment_async.py
"""
Async MixRank Enrichment Module

Parallel API calls using asyncio + aiohttp for high throughput.
Achieves 10-20+ req/sec vs ~1.5 req/sec sequential.
"""

import asyncio
import os
import time
from dataclasses import dataclass
from typing import Any, Optional

import aiohttp
from dotenv import load_dotenv

load_dotenv()


@dataclass
class EnrichmentResult:
    """Result of a single enrichment API call."""
    record_id: str
    linkedin_url: str
    success: bool
    data: Optional[dict] = None
    error: Optional[str] = None
    http_status: Optional[int] = None


class AsyncMixRankEnricher:
    """
    Async MixRank API client for high-throughput LinkedIn profile enrichment.
    
    Features:
    - Concurrent requests (configurable concurrency)
    - Semaphore-based rate limiting
    - Automatic retry with backoff
    - Batch processing
    """
    
    def __init__(
        self,
        api_key: Optional[str] = None,
        concurrency: int = 20,
        timeout: int = 30,
        max_retries: int = 3
    ):
        self.api_key = api_key or os.environ.get("MIXRANK_API_KEY")
        if not self.api_key:
            raise ValueError(
                "MixRank API key required. Set MIXRANK_API_KEY env var or pass api_key parameter."
            )
        
        self.base_url = f"https://api.mixrank.com/v2/json/{self.api_key}/linkedin/profile"
        self.concurrency = concurrency
        self.timeout = aiohttp.ClientTimeout(total=timeout)
        self.max_retries = max_retries
        
        # Semaphore to limit concurrent requests
        self._semaphore: Optional[asyncio.Semaphore] = None
        
        # Statistics
        self.stats = {
            "total_requests": 0,
            "successful": 0,
            "failed": 0,
            "retries": 0,
        }
    
    async def _ensure_semaphore(self):
        """Create semaphore if not exists (must be created in async context)."""
        if self._semaphore is None:
            self._semaphore = asyncio.Semaphore(self.concurrency)
    
    async def enrich_profile(
        self,
        session: aiohttp.ClientSession,
        record_id: str,
        linkedin_url: str
    ) -> EnrichmentResult:
        """
        Enrich a single LinkedIn profile with retry logic.
        """
        await self._ensure_semaphore()
        
        self.stats["total_requests"] += 1
        
        async with self._semaphore:
            for attempt in range(self.max_retries + 1):
                try:
                    async with session.get(
                        self.base_url,
                        params={"url": linkedin_url},
                        timeout=self.timeout
                    ) as response:
                        
                        if response.status == 200:
                            data = await response.json()
                            self.stats["successful"] += 1
                            return EnrichmentResult(
                                record_id=record_id,
                                linkedin_url=linkedin_url,
                                success=True,
                                data=data,
                                http_status=200
                            )
                        
                        # Rate limited - wait and retry
                        if response.status == 429:
                            if attempt < self.max_retries:
                                self.stats["retries"] += 1
                                wait_time = 2 ** attempt  # Exponential backoff
                                await asyncio.sleep(wait_time)
                                continue
                        
                        # Server error - retry
                        if response.status >= 500:
                            if attempt < self.max_retries:
                                self.stats["retries"] += 1
                                await asyncio.sleep(1)
                                continue
                        
                        # Client error or max retries reached
                        text = await response.text()
                        self.stats["failed"] += 1
                        return EnrichmentResult(
                            record_id=record_id,
                            linkedin_url=linkedin_url,
                            success=False,
                            error=f"HTTP {response.status}: {text[:200]}",
                            http_status=response.status
                        )
                        
                except asyncio.TimeoutError:
                    if attempt < self.max_retries:
                        self.stats["retries"] += 1
                        continue
                    self.stats["failed"] += 1
                    return EnrichmentResult(
                        record_id=record_id,
                        linkedin_url=linkedin_url,
                        success=False,
                        error="Request timeout"
                    )
                    
                except Exception as e:
                    self.stats["failed"] += 1
                    return EnrichmentResult(
                        record_id=record_id,
                        linkedin_url=linkedin_url,
                        success=False,
                        error=f"Request error: {str(e)}"
                    )
        
        # Should not reach here
        self.stats["failed"] += 1
        return EnrichmentResult(
            record_id=record_id,
            linkedin_url=linkedin_url,
            success=False,
            error="Unknown error"
        )
    
    async def enrich_batch(
        self,
        records: list[tuple[str, str]],  # List of (record_id, linkedin_url)
        progress_callback: Optional[callable] = None
    ) -> list[EnrichmentResult]:
        """
        Enrich a batch of profiles concurrently.
        
        Args:
            records: List of (record_id, linkedin_url) tuples
            progress_callback: Optional callback(result) called after each completion
        
        Returns:
            List of EnrichmentResult in same order as input
        """
        results = []
        
        async with aiohttp.ClientSession() as session:
            # Create tasks for all records
            tasks = [
                self.enrich_profile(session, record_id, url)
                for record_id, url in records
            ]
            
            # Process as they complete
            for coro in asyncio.as_completed(tasks):
                result = await coro
                results.append(result)
                
                if progress_callback:
                    progress_callback(result)
        
        return results
    
    def get_stats(self) -> dict:
        """Get current statistics."""
        return self.stats.copy()


class AsyncProgressTracker:
    """
    Track and display progress for async operations.
    Thread-safe for concurrent updates.
    """
    
    def __init__(
        self,
        total: int,
        description: str = "Processing",
        update_interval: int = 10
    ):
        self.total = total
        self.description = description
        self.update_interval = update_interval
        
        self.processed = 0
        self.success = 0
        self.errors = 0
        
        self.start_time = time.time()
        self._lock = asyncio.Lock()
    
    async def update(self, success: bool = True):
        """Update progress (thread-safe)."""
        async with self._lock:
            self.processed += 1
            if success:
                self.success += 1
            else:
                self.errors += 1
            
            if self.processed % self.update_interval == 0 or self.processed == self.total:
                self._print_progress()
    
    def _print_progress(self):
        """Print progress bar."""
        elapsed = time.time() - self.start_time
        pct = (self.processed / self.total) * 100 if self.total > 0 else 0
        
        if self.processed > 0:
            rate = self.processed / elapsed
            remaining = self.total - self.processed
            eta_seconds = remaining / rate if rate > 0 else 0
            eta_str = self._format_time(eta_seconds)
        else:
            rate = 0
            eta_str = "--:--"
        
        bar_width = 40
        filled = int(bar_width * self.processed / self.total) if self.total > 0 else 0
        bar = "█" * filled + "░" * (bar_width - filled)
        
        print(
            f"\r{self.description}: {pct:5.1f}% |{bar}| "
            f"{self.processed}/{self.total} [{self._format_time(elapsed)}<{eta_str}] "
            f"✓{self.success} ✗{self.errors} ({rate:.1f}/s)",
            end="",
            flush=True
        )
        
        if self.processed == self.total:
            print()
    
    def _format_time(self, seconds: float) -> str:
        if seconds < 0:
            return "--:--"
        hours = int(seconds // 3600)
        minutes = int((seconds % 3600) // 60)
        secs = int(seconds % 60)
        if hours > 0:
            return f"{hours}:{minutes:02d}:{secs:02d}"
        return f"{minutes:02d}:{secs:02d}"
    
    def finish(self):
        """Print final summary."""
        elapsed = time.time() - self.start_time
        rate = self.processed / elapsed if elapsed > 0 else 0
        print()
        print("=" * 60)
        print(f"  {self.description} Complete!")
        print(f"  Total: {self.processed} | Success: {self.success} | Errors: {self.errors}")
        print(f"  Time: {self._format_time(elapsed)} | Rate: {rate:.1f}/sec")
        print("=" * 60)


# Testing
if __name__ == "__main__":
    import sys
    
    async def test_async_enricher():
        print("Async Enrichment Module Test")
        print("=" * 50)
        
        # Check for API key
        if not os.environ.get("MIXRANK_API_KEY"):
            print("⚠ MIXRANK_API_KEY not set, skipping live test")
            return
        
        # Test with a few URLs
        test_records = [
            ("test_1", "https://www.linkedin.com/in/satya-nadella-3145136"),
            ("test_2", "https://www.linkedin.com/in/sundarpichai"),
            ("test_3", "https://www.linkedin.com/in/timcook"),
        ]
        
        enricher = AsyncMixRankEnricher(concurrency=5)
        
        tracker = AsyncProgressTracker(
            total=len(test_records),
            description="Test Enrichment",
            update_interval=1
        )
        
        async def on_result(result):
            await tracker.update(success=result.success)
        
        print(f"\nTesting with {len(test_records)} profiles...")
        results = await enricher.enrich_batch(test_records, progress_callback=on_result)
        tracker.finish()
        
        print(f"\nResults:")
        for r in results:
            status = "✓" if r.success else "✗"
            print(f"  {status} {r.record_id}: {r.linkedin_url[:40]}...")
            if r.error:
                print(f"      Error: {r.error}")
        
        print(f"\nStats: {enricher.get_stats()}")
    
    asyncio.run(test_async_enricher())

