requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.9.0

